migrate Python code to a dedicated package
