+++ src/setup.cfg
@@ -1,0 +1,2 @@
+[metadata]
+version = attr: tesstrain.__version__
